+++ software/deimos/src/controller/context.rs
@@ -1,13 +1,17 @@
 //! Information about the current operation
 //! that may be used by the controller's appendages.
 
+use std::ops::Deref;
 use std::path::PathBuf;
+use std::sync::{Arc, RwLock};
 use std::time::{Duration, SystemTime};
 use std::{collections::BTreeMap, default::Default};
 
 use chrono::{DateTime, Utc};
 
 use serde::{Deserialize, Serialize};
+
+use super::channel::{Channel, Endpoint};
 
 /// Criteria for exiting the control program
 #[derive(Serialize, Deserialize, Clone, Debug)]
@@ -84,9 +88,40 @@ pub struct ControllerCtx {
     /// Response to losing contact with a peripheral
     pub loss_of_contact_policy: LossOfContactPolicy,
 
-    /// A last-resort escape hatch for sideloading user context (likely json-encoded)
+    /// An escape hatch for sideloading user context (likely json-encoded)
     /// that is not yet implemented as a standalone field.
     pub user_ctx: BTreeMap<String, String>,
+
+    /// An escape hatch for sideloading communication between appendages.
+    /// Each channel is a bidirectional MPMC message pipe.
+    ///
+    /// Because bidirectional channels may not close until the program terminates
+    /// on its own, the status of these channels should not be used to indicate
+    /// when a freerunning thread should terminate, as this will often result in
+    /// a resource leak.
+    pub user_channels: Arc<RwLock<BTreeMap<String, Channel>>>,
+}
+
+impl ControllerCtx {
+    /// Get a handle to a source endpoint tx/rx pair for the channel,
+    /// creating the channel if it does not exist.
+    pub fn source_endpoint(&self, channel_name: &str) -> Endpoint {
+        let map = &self.user_channels;
+        let inner = map.deref();
+        let mut writer = inner.try_write().unwrap();
+        let channel = writer.entry(channel_name.to_owned()).or_default();
+        channel.source_endpoint()
+    }
+
+    /// Get a handle to a sink endpoint tx/rx pair for the channel,
+    /// creating the channel if it does not exist.
+    pub fn sink_endpoint(&self, channel_name: &str) -> Endpoint {
+        let map = &self.user_channels;
+        let inner = map.deref();
+        let mut writer = inner.try_write().unwrap();
+        let channel = writer.entry(channel_name.to_owned()).or_default();
+        channel.sink_endpoint()
+    }
 }
 
 impl Default for ControllerCtx {
@@ -108,6 +143,7 @@ impl Default for ControllerCtx {
             termination_criteria: Vec::new(),
             loss_of_contact_policy: LossOfContactPolicy::Terminate,
             user_ctx: BTreeMap::new(),
+            user_channels: Arc::new(RwLock::new(BTreeMap::new())),
         }
     }
 }